laird-linux-backports: rework recipe
This rework was required to fix a build issue seen on build hosts where
depomod is not available. This was evidence of some host tool contamination.

The following changes do address this and while at it clean up the recipe
a bit:

* add patch to prevent upstream makefiles from running module post installation
steps, namely depmod and update-initramfs.sh. This is already carried out by
module.bbclass this recipe inherits (depmod) or not relevant for the OE build
system (update-initramfs.sh)
* move RDEPENDS_${PN} to the end of the recipe
* use already set PV to define BRANCH variable (avoid hardcoding version twice)
* set S appropritely
* use EXTRA_OEMAKE for passing KLIB_BUILD and KLIB to upstream Makefile
* do_compile(): drop custom version
* do_configure(): use oe_runmake and set host compiler appropriately
* do_install(): use the task provided by module class